requirements.txt
openai
pdfplumber
pandas
python-dotenv